Worked on the blockchain application

+++ roots/applications/blockchains/sdk.go
@@ -58,7 +58,7 @@ type Blockchain interface {
 
 // Reference represents the reference application
 type Reference interface {
-	ContentKeys(context uint, kind uint8) (references.ContentKeys, error)
+	ContentKeys(context uint, kind uint) (references.ContentKeys, error)
 	ContentKey(context uint, hash hash.Hash, flag uint8) (references.ContentKey, error)
 	ContentKeyByTransaction(context uint, trx hash.Hash, flag uint8) (references.ContentKey, error)
 	BlockchainKey(context uint, hash hash.Hash, flag uint8) (references.BlockchainKey, error)
@@ -80,5 +80,5 @@ type Content interface {
 	ReadByHash(content uint, hash hash.Hash) ([]byte, error)
 	ReadAll(context uint, pointers []references.Pointer) ([][]byte, error)
 	ReadAllByHashes(context uint, hashes []hash.Hash) ([][]byte, error)
-	Write(context uint, hash hash.Hash, data []byte, kind uint8) error
+	Write(context uint, hash hash.Hash, data []byte, kind uint) error
 }
